requirements.txt
future==0.17.1
future-stubs==0.18.0.0
google-pasta==0.1.8
grpcio==1.20.1
kiwisolver==1.1.0
lpdn==0.0.2
lucid==0.3.8
opencv-contrib-python==4.1.2.30
opencv-python==4.1.0.25
tensorboard==1.13.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.2
tensorflow-probability==0.7.0

